Rename PersistentStorage packages
