--- .github/workflows/package-Extensions.PersistentStorage.yml
@@ -1,67 +1,0 @@
-name: Pack OpenTelemetry.Extensions.PersistentStorage
-
-on:
-  workflow_dispatch:
-    inputs:
-      logLevel:
-        description: 'Log level'
-        required: true
-        default: 'warning'
-  push:
-    tags:
-      - 'Extensions.PersistentStorage-*' # trigger when we create a tag with prefix "Extensions.PersistentStorage-"
-
-jobs:
-  build-test-pack:
-    runs-on: ${{ matrix.os }}
-    permissions:
-      contents: write
-    env:
-      PROJECT: OpenTelemetry.Extensions.PersistentStorage
-
-    strategy:
-      matrix:
-        os: [windows-latest]
-
-    steps:
-    - uses: actions/checkout@v3
-      with:
-        fetch-depth: 0 # fetching all
-
-    - uses: actions/setup-dotnet@v3.0.3
-      with:
-        dotnet-version: '7.0.x'
-
-    - name: Install dependencies
-      run: dotnet restore src/${{env.PROJECT}}
-
-    - name: dotnet build ${{env.PROJECT}}
-      run: dotnet build src/${{env.PROJECT}} --configuration Release --no-restore -p:Deterministic=true
-
-    - name: dotnet test ${{env.PROJECT}}
-      run: dotnet test test/${{env.PROJECT}}.Tests
-
-    - name: dotnet pack ${{env.PROJECT}}
-      run: dotnet pack src/${{env.PROJECT}} --configuration Release #--no-build <- OpenTelemetry.Extensions.PersistentStorage has a conditional net6.0 target which causes dotnet pack to break when -no-build is used (for some reason)
-
-    - name: Publish Artifacts
-      uses: actions/upload-artifact@v3
-      with:
-        name: ${{env.PROJECT}}-packages
-        path: '**/${{env.PROJECT}}/bin/**/*.*nupkg'
-
-    - name: Publish Nuget
-      run: |
-        nuget push **/${{env.PROJECT}}/bin/**/*.nupkg -Source https://api.nuget.org/v3/index.json -ApiKey ${{ secrets.NUGET_TOKEN }} -SymbolApiKey ${{ secrets.NUGET_TOKEN }}
-
-    - name: Create GitHub Prerelease
-      if: ${{ (contains(github.ref_name, '-alpha.') || contains(github.ref_name, '-beta.') || contains(github.ref_name, '-rc.')) }}
-      run: gh release create ${{ github.ref_name }} --title ${{ github.ref_name }} --verify-tag --notes "See [CHANGELOG](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/blob/${{ github.ref_name }}/src/${{env.PROJECT}}/CHANGELOG.md) for details." --prerelease
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-    - name: Create GitHub Release
-      if: ${{ !(contains(github.ref_name, '-alpha.') || contains(github.ref_name, '-beta.') || contains(github.ref_name, '-rc.')) }}
-      run: gh release create ${{ github.ref_name }} --title ${{ github.ref_name }} --verify-tag --notes "See [CHANGELOG](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/blob/${{ github.ref_name }}/src/${{env.PROJECT}}/CHANGELOG.md) for details." --latest
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
