--- src/OpenTelemetry.Extensions.PersistentStorage.Abstractions/.publicApi/netstandard2.0/PublicAPI.Unshipped.txt
@@ -1,20 +1,0 @@
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.OnTryDelete() -> bool
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.OnTryLease(int leasePeriodMilliseconds) -> bool
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.OnTryRead(out byte[]? buffer) -> bool
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.OnTryWrite(byte[]! buffer, int leasePeriodMilliseconds = 0) -> bool
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.OnGetBlobs() -> System.Collections.Generic.IEnumerable<OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob!>!
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.OnTryCreateBlob(byte[]! buffer, int leasePeriodMilliseconds, out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob? blob) -> bool
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.OnTryCreateBlob(byte[]! buffer, out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob? blob) -> bool
-abstract OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.OnTryGetBlob(out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob? blob) -> bool
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.PersistentBlob() -> void
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.TryDelete() -> bool
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.TryLease(int leasePeriodMilliseconds) -> bool
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.TryRead(out byte[]? buffer) -> bool
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob.TryWrite(byte[]! buffer, int leasePeriodMilliseconds = 0) -> bool
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.GetBlobs() -> System.Collections.Generic.IEnumerable<OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob!>!
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.PersistentBlobProvider() -> void
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.TryCreateBlob(byte[]! buffer, int leasePeriodMilliseconds, out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob? blob) -> bool
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.TryCreateBlob(byte[]! buffer, out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob? blob) -> bool
-OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlobProvider.TryGetBlob(out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob? blob) -> bool
