--- src/OpenTelemetry.Extensions.PersistentStorage.Abstractions/CHANGELOG.md
@@ -1,11 +1,0 @@
-# Changelog
-
-## 1.0.0-beta.1
-
-## 1.0.0-alpha.4
-
-This is the first release for the `OpenTelemetry.Extensions.PersistentStorage.Abstractions`
-project.
-
-For more details, please refer to the
-[README](README.md)
