--- src/OpenTelemetry.Extensions.PersistentStorage.Abstractions/README.md
@@ -1,13 +1,0 @@
-# Persistent Storage Abstractions
-
-[![NuGet](https://img.shields.io/nuget/v/OpenTelemetry.Extensions.PersistentStorage.Abstractions.svg)](https://www.nuget.org/packages/OpenTelemetry.Extensions.PersistentStorage.Abstractions)
-[![NuGet](https://img.shields.io/nuget/dt/OpenTelemetry.Extensions.PersistentStorage.Abstractions.svg)](https://www.nuget.org/packages/OpenTelemetry.Extensions.PersistentStorage.Abstractions)
-
-This package includes APIs which can be extended by exporter owners to implement
-persistent storage.
-
-## Installation
-
-```shell
-dotnet add package OpenTelemetry.Extensions.PersistentStorage.Abstractions
-```
