--- src/OpenTelemetry.Extensions.PersistentStorage/.publicApi/net462/PublicAPI.Unshipped.txt
@@ -1,15 +1,0 @@
-OpenTelemetry.Extensions.PersistentStorage.FileBlob
-OpenTelemetry.Extensions.PersistentStorage.FileBlob.FileBlob(string fullPath) -> void
-OpenTelemetry.Extensions.PersistentStorage.FileBlob.FullPath.get -> string
-OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider
-OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider.Dispose() -> void
-OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider.FileBlobProvider(string path, long maxSizeInBytes = 52428800, int maintenancePeriodInMilliseconds = 120000, long retentionPeriodInMilliseconds = 172800000, int writeTimeoutInMilliseconds = 60000) -> void
-override OpenTelemetry.Extensions.PersistentStorage.FileBlob.OnTryDelete() -> bool
-override OpenTelemetry.Extensions.PersistentStorage.FileBlob.OnTryLease(int leasePeriodMilliseconds) -> bool
-override OpenTelemetry.Extensions.PersistentStorage.FileBlob.OnTryRead(out byte[] buffer) -> bool
-override OpenTelemetry.Extensions.PersistentStorage.FileBlob.OnTryWrite(byte[] buffer, int leasePeriodMilliseconds = 0) -> bool
-override OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider.OnGetBlobs() -> System.Collections.Generic.IEnumerable<OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob>
-override OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider.OnTryCreateBlob(byte[] buffer, int leasePeriodMilliseconds, out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob blob) -> bool
-override OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider.OnTryCreateBlob(byte[] buffer, out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob blob) -> bool
-override OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider.OnTryGetBlob(out OpenTelemetry.Extensions.PersistentStorage.Abstractions.PersistentBlob blob) -> bool
-virtual OpenTelemetry.Extensions.PersistentStorage.FileBlobProvider.Dispose(bool disposing) -> void
