--- src/OpenTelemetry.Extensions.PersistentStorage/OpenTelemetry.Extensions.PersistentStorage.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <!-- OmniSharp/VS Code requires TargetFrameworks to be in descending order for IntelliSense and analysis. -->
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <TargetFrameworks Condition="$(OS) == 'Windows_NT'">$(TargetFrameworks);net462</TargetFrameworks>
-    <TargetFrameworks Condition="'$(Configuration)' == 'Debug'">net6.0;$(TargetFrameworks)</TargetFrameworks> <!-- Added just to get proper nullable analysis in IDE -->
-    <Description>OpenTelemetry Persistent Storage</Description>
-    <MinVerTagPrefix>Extensions.PersistentStorage-</MinVerTagPrefix>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <NoWarn>$(NoWarn),1591</NoWarn>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="$(RepoRoot)\src\OpenTelemetry.Internal\Guard.cs" Link="Includes\Guard.cs" />
-    <Compile Include="$(RepoRoot)\src\OpenTelemetry.Internal\NullableAttributes.cs" Link="Includes\NullableAttributes.cs" />
-    <ProjectReference Include="$(RepoRoot)\src\OpenTelemetry.Extensions.PersistentStorage.Abstractions\OpenTelemetry.Extensions.PersistentStorage.Abstractions.csproj" />
-  </ItemGroup>
-
-</Project>
