--- src/OpenTelemetry.Extensions.PersistentStorage/PersistentStorageHelper.cs
@@ -1,259 +1,0 @@
-// <copyright file="PersistentStorageHelper.cs" company="OpenTelemetry Authors">
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// </copyright>
-
-using System;
-using System.Diagnostics;
-using System.Globalization;
-using System.IO;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Security.Cryptography;
-using System.Text;
-using System.Threading;
-
-namespace OpenTelemetry.Extensions.PersistentStorage;
-
-internal static class PersistentStorageHelper
-{
-    private static long directorySize;
-
-    internal static void RemoveExpiredBlob(DateTime retentionDeadline, string filePath)
-    {
-        if (filePath.EndsWith(".blob", StringComparison.OrdinalIgnoreCase))
-        {
-            DateTime fileDateTime = GetDateTimeFromBlobName(filePath);
-            if (fileDateTime < retentionDeadline)
-            {
-                try
-                {
-                    File.Delete(filePath);
-                    PersistentStorageEventSource.Log.PersistentStorageInformation(nameof(PersistentStorageHelper), "Removing blob as retention deadline expired");
-                }
-                catch (Exception ex)
-                {
-                    PersistentStorageEventSource.Log.CouldNotRemoveExpiredBlob(filePath, ex);
-                }
-            }
-        }
-    }
-
-    internal static bool RemoveExpiredLease(DateTime leaseDeadline, string filePath)
-    {
-        bool success = false;
-
-        if (filePath.EndsWith(".lock", StringComparison.OrdinalIgnoreCase))
-        {
-            DateTime fileDateTime = GetDateTimeFromLeaseName(filePath);
-            if (fileDateTime < leaseDeadline)
-            {
-                var newFilePath = filePath.Substring(0, filePath.LastIndexOf('@'));
-                try
-                {
-                    File.Move(filePath, newFilePath);
-                    success = true;
-                }
-                catch (Exception ex)
-                {
-                    PersistentStorageEventSource.Log.CouldNotRemoveExpiredLease(filePath, newFilePath, ex);
-                }
-            }
-        }
-
-        return success;
-    }
-
-    internal static bool RemoveTimedOutTmpFiles(DateTime timeoutDeadline, string filePath)
-    {
-        bool success = false;
-
-        if (filePath.EndsWith(".tmp", StringComparison.OrdinalIgnoreCase))
-        {
-            DateTime fileDateTime = GetDateTimeFromBlobName(filePath);
-            if (fileDateTime < timeoutDeadline)
-            {
-                try
-                {
-                    File.Delete(filePath);
-                    success = true;
-                    PersistentStorageEventSource.Log.PersistentStorageInformation(nameof(PersistentStorageHelper), "File write exceeded timeout. Dropping telemetry");
-                }
-                catch (Exception ex)
-                {
-                    PersistentStorageEventSource.Log.CouldNotRemoveTimedOutTmpFile(filePath, ex);
-                }
-            }
-        }
-
-        return success;
-    }
-
-    internal static void RemoveExpiredBlobs(string directoryPath, long retentionPeriodInMilliseconds, long writeTimeoutInMilliseconds)
-    {
-        var currentUtcDateTime = DateTime.UtcNow;
-
-        var leaseDeadline = currentUtcDateTime;
-        var retentionDeadline = currentUtcDateTime - TimeSpan.FromMilliseconds(retentionPeriodInMilliseconds);
-        var timeoutDeadline = currentUtcDateTime - TimeSpan.FromMilliseconds(writeTimeoutInMilliseconds);
-
-        foreach (var file in Directory.EnumerateFiles(directoryPath).OrderByDescending(filename => filename))
-        {
-            var success = RemoveTimedOutTmpFiles(timeoutDeadline, file);
-
-            if (success)
-            {
-                continue;
-            }
-
-            success = RemoveExpiredLease(leaseDeadline, file);
-
-            if (!success)
-            {
-                RemoveExpiredBlob(retentionDeadline, file);
-            }
-        }
-
-        // It is faster to calculate the directory size, instead of removing length of expired files.
-        var size = CalculateFolderSize(directoryPath);
-        Interlocked.Exchange(ref directorySize, size);
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static long GetDirectorySize()
-    {
-        return Interlocked.Read(ref directorySize);
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static void WriteAllBytes(string path, byte[] buffer)
-    {
-        File.WriteAllBytes(path, buffer);
-        UpdateDirectorySize(buffer.Length);
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static void RemoveFile(string fileName)
-    {
-        var fileInfo = new FileInfo(fileName);
-        var fileSize = fileInfo.Length;
-        fileInfo.Delete();
-        UpdateDirectorySize(fileSize * -1);
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static string GetUniqueFileName(string extension)
-    {
-        return string.Format(CultureInfo.InvariantCulture, $"{DateTime.UtcNow:yyyy-MM-ddTHHmmss.fffffffZ}-{Guid.NewGuid().ToString("N", CultureInfo.InvariantCulture)}{extension}");
-    }
-
-    internal static string CreateSubdirectory(string path)
-    {
-        string subdirectoryPath = string.Empty;
-        string baseDirectory = string.Empty;
-#if !NETSTANDARD
-        baseDirectory = AppDomain.CurrentDomain.BaseDirectory;
-#else
-        baseDirectory = AppContext.BaseDirectory;
-#endif
-
-        string appIdentity = Environment.UserName + "@" + Path.Combine(baseDirectory, Process.GetCurrentProcess().ProcessName);
-        string subdirectoryName = GetSHA256Hash(appIdentity);
-        subdirectoryPath = Path.Combine(path, subdirectoryName);
-        Directory.CreateDirectory(subdirectoryPath);
-
-        directorySize = CalculateFolderSize(subdirectoryPath);
-        return subdirectoryPath;
-    }
-
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static long UpdateDirectorySize(long fileContentLength)
-    {
-        return Interlocked.Add(ref directorySize, fileContentLength);
-    }
-
-    internal static DateTime GetDateTimeFromBlobName(string filePath)
-    {
-        var fileName = Path.GetFileNameWithoutExtension(filePath);
-        var time = fileName.Substring(0, fileName.LastIndexOf('-'));
-
-        // TODO:Handle possible parsing failure.
-        DateTime.TryParseExact(time, "yyyy-MM-ddTHHmmss.fffffffZ", CultureInfo.InvariantCulture, DateTimeStyles.None, out var dateTime);
-        return dateTime.ToUniversalTime();
-    }
-
-    internal static DateTime GetDateTimeFromLeaseName(string filePath)
-    {
-        var fileName = Path.GetFileNameWithoutExtension(filePath);
-        var startIndex = fileName.LastIndexOf('@') + 1;
-        var time = fileName.Substring(startIndex, fileName.Length - startIndex);
-        DateTime.TryParseExact(time, "yyyy-MM-ddTHHmmss.fffffffZ", CultureInfo.InvariantCulture, DateTimeStyles.None, out var dateTime);
-        return dateTime.ToUniversalTime();
-    }
-
-    internal static string GetSHA256Hash(string input)
-    {
-        byte[] inputBits = Encoding.Unicode.GetBytes(input);
-
-#if NET6_0_OR_GREATER
-#pragma warning disable CA1308 // Normalize strings to uppercase
-        return Convert.ToHexString(SHA256.HashData(inputBits)).ToLowerInvariant();
-#pragma warning restore CA1308 // Normalize strings to uppercase
-#else
-        var hashString = new StringBuilder();
-        using (var sha256 = SHA256.Create())
-        {
-            byte[] hashBits = sha256.ComputeHash(inputBits);
-            foreach (byte b in hashBits)
-            {
-                hashString.Append(b.ToString("x2", CultureInfo.InvariantCulture));
-            }
-        }
-
-        return hashString.ToString();
-#endif
-    }
-
-    private static long CalculateFolderSize(string path)
-    {
-        if (!Directory.Exists(path))
-        {
-            return 0;
-        }
-
-        long directorySize = 0;
-        try
-        {
-            foreach (string file in Directory.EnumerateFiles(path))
-            {
-                if (File.Exists(file))
-                {
-                    FileInfo fileInfo = new FileInfo(file);
-                    directorySize += fileInfo.Length;
-                }
-            }
-
-            foreach (string dir in Directory.GetDirectories(path))
-            {
-                directorySize += CalculateFolderSize(dir);
-            }
-        }
-        catch (Exception ex)
-        {
-            PersistentStorageEventSource.Log.PersistentStorageException(nameof(PersistentStorageHelper), "Error calculating folder size", ex);
-        }
-
-        return directorySize;
-    }
-}
