+++ src/OpenTelemetry.PersistentStorage.Abstractions/.publicApi/net462/PublicAPI.Shipped.txt
