+++ src/OpenTelemetry.PersistentStorage.Abstractions/.publicApi/net6.0/PublicAPI.Shipped.txt
