+++ src/OpenTelemetry.PersistentStorage.Abstractions/.publicApi/netstandard2.0/PublicAPI.Shipped.txt
