+++ src/OpenTelemetry.PersistentStorage.Abstractions/AssemblyInfo.cs
