+++ src/OpenTelemetry.PersistentStorage.Abstractions/CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+
+// TODO: update changelog
+
+## 1.0.0-beta.1
+
+## 1.0.0-alpha.4
+
+This is the first release for the `OpenTelemetry.Extensions.PersistentStorage.Abstractions`
+project.
+
+For more details, please refer to the
+[README](README.md)
